Rename icons file
