+++ lib/ui/common_widgets/bookmark_button.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:patta/local_database/database.dart';
 import 'package:patta/local_database/moor_converters.dart' as moor_converters;
-import 'package:patta/ui/common_widgets/icons.dart';
+import 'package:patta/ui/common_widgets/pariyatti_icons.dart';
 import 'package:patta/ui/model/CardModel.dart';
 
 class BookmarkButton extends StatelessWidget {
