+++ lib/ui/common_widgets/cards/StackedInspirationCard.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 import 'package:patta/local_database/database.dart';
 import 'package:patta/resources/strings.dart';
 import 'package:patta/ui/common_widgets/bookmark_button.dart';
-import 'package:patta/ui/common_widgets/icons.dart';
+import 'package:patta/ui/common_widgets/pariyatti_icons.dart';
 import 'package:patta/ui/common_widgets/share_button.dart';
 import 'package:patta/ui/model/StackedInspirationCardModel.dart';
 import 'package:patta/util.dart';
