+++ lib/ui/common_widgets/share_button.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:patta/ui/common_widgets/icons.dart';
+import 'package:patta/ui/common_widgets/pariyatti_icons.dart';
 
 class ShareButton extends StatelessWidget {
   Function onPressed;
